+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Most recent change on the bottom.
 ## [Unreleased] - 0.5.6
 ### Added
 - `nequip-benchmark` and `nequip-train` report number of weights and number of trainable weights
-- `nequip-benchmark --no-compile` and `--verbose`
+- `nequip-benchmark --no-compile` and `--verbose` and `--memory-summary`
 
 ### Fixed
 - Git utilities when installed as ZIPed `.egg` (#264)